improved tests folder structure + centralized fixtures for subfolder
