+++ api/tests/exceptions/test_exception_handlers.py
@@ -1,14 +1,3 @@
-import pytest
-from botocore.exceptions import ClientError
-
-
-@pytest.fixture()
-def client_error_response():
-    error_response = dict(Error={'Code': 400, 'Message': 'Operation failed'})
-    error_response['ResponseMetadata'] = dict(HTTPStatusCode=400)
-    return ClientError(error_response, 'failed_operation')
-
-
 def test_boto3_exception_handler(client, client_error_response, app, monkeypatch):
     def delete_user_raising_clienterror():
         raise client_error_response
